Corrected the main program:
1.  The diffuse and the direct  radiation in the weather data file are
swapped, i.e. what was read as  direct was diffuse and vice versa. The
columns in the weather data file are now correctly interpreted.

2.   The sun  path is  wrong  in the  weather data  file. The  azimuth
increases up  till the noon but  then decreases. This  is corrected in
the program and values in the weather data file are not touched.

3. The sky  (Firmament) is reconfigured with the  value of the current
value of diffuse  light, the direct radiation is set  as before. We do
need the value of the diffuse light on the light model.

At the end of the short time step the diffuse radiation is set to last
value  read, but the  last value  is 0  (or should  be, because  it is
night).